refactor: consolidate EditMatchForm and improve player management
- Moved EditMatchForm to a new component file for better organization.
- Updated player handling logic to include type distinction for players.
- Enhanced form validation and submission process for improved user experience.
- Streamlined location management within the match editing process.
- Removed deprecated components and optimized imports for clarity and maintainability.
